make integral types consistent

+++ include/menu.h
@@ -6,7 +6,7 @@
 #pragma once
 
 #include "raylib.h"
-#include <stdlib.h>
+#include <stdint.h>
 
 /**
  * @brief Individual menu item structure
@@ -16,7 +16,7 @@ typedef struct
     char *text;           ///< Text to display for this menu item
     int pos_x;            ///< X position for rendering
     int pos_y;            ///< Y position for rendering
-    int font_size;        ///< Font size for text rendering
+    uint16_t font_size;   ///< Font size for text rendering
     Color color;          ///< Text color
     void (*select)(void); ///< Function to call when item is selected
 } MenuItem;
@@ -26,10 +26,10 @@ typedef struct
  */
 typedef struct
 {
-    size_t num_rows;    ///< Number of rows in the grid
-    size_t num_cols;    ///< Number of columns in the grid
-    size_t current_row; ///< Currently selected row
-    size_t current_col; ///< Currently selected column
+    uint16_t num_rows;    ///< Number of rows in the grid
+    uint16_t num_cols;    ///< Number of columns in the grid
+    uint16_t current_row; ///< Currently selected row
+    uint16_t current_col; ///< Currently selected column
 } Grid;
 
 /**
@@ -42,7 +42,7 @@ typedef struct
     void (*move_down)(const void *);  ///< Function to handle down movement
     void (*move_up)(const void *);    ///< Function to handle up movement
     Grid grid;                        ///< Grid navigation state
-    size_t num_items;                 ///< Total number of menu items
+    uint16_t num_items;               ///< Total number of menu items
     MenuItem *items[];                ///< Array of menu item pointers
 } GridMenu;
 
@@ -54,8 +54,8 @@ typedef struct
     void (*next_item)(const void *); ///< Function to move to next item
     void (*prev_item)(const void *); ///< Function to move to previous item
     void (*select)(const void *);    ///< Function to select current item
-    size_t selected_item;            ///< Index of currently selected item
-    size_t num_items;                ///< Total number of menu items
+    uint16_t selected_item;          ///< Index of currently selected item
+    uint16_t num_items;              ///< Total number of menu items
     MenuItem items[];                ///< Array of menu items
 } VerticalMenu;
 
@@ -67,7 +67,8 @@ typedef struct
  * @param num_cols Number of columns in the grid
  * @return Pointer to the created grid menu
  */
-GridMenu *grid_menu_create(const size_t num_items, const size_t num_rows, const size_t num_cols);
+GridMenu *grid_menu_create(const uint16_t num_items, const uint16_t num_rows,
+                           const uint16_t num_cols);
 
 /**
  * @brief Destroys a grid menu and frees its memory
@@ -82,7 +83,7 @@ void grid_menu_destroy(GridMenu *menu);
  * @param num_items Number of menu items
  * @return Pointer to the created vertical menu
  */
-VerticalMenu *vertical_menu_create(const size_t num_items);
+VerticalMenu *vertical_menu_create(const uint16_t num_items);
 
 /**
  * @brief Destroys a vertical menu and frees its memory
